Verify logic is added to userCreate API operation
