+++ src/main/java/gov/ca/cwds/idm/service/validation/ValidationServiceImpl.java
@@ -1,35 +1,53 @@
 package gov.ca.cwds.idm.service.validation;
 
+import static gov.ca.cwds.config.api.idm.Roles.COUNTY_ADMIN;
+import static gov.ca.cwds.config.api.idm.Roles.OFFICE_ADMIN;
+import static gov.ca.cwds.idm.service.authorization.UserRolesService.getStrongestAdminRole;
+import static gov.ca.cwds.idm.service.cognito.util.CognitoUsersSearchCriteriaUtil.composeToGetFirstPageByEmail;
 import static gov.ca.cwds.idm.service.cognito.util.CognitoUsersSearchCriteriaUtil.composeToGetFirstPageByRacfId;
 import static gov.ca.cwds.service.messages.MessageCode.ACTIVE_USER_WITH_RAFCID_EXISTS_IN_IDM;
+import static gov.ca.cwds.service.messages.MessageCode.NOT_AUTHORIZED_TO_ADD_USER_FOR_OTHER_COUNTY;
+import static gov.ca.cwds.service.messages.MessageCode.NOT_AUTHORIZED_TO_ADD_USER_FOR_OTHER_OFFICE;
 import static gov.ca.cwds.service.messages.MessageCode.NO_USER_WITH_RACFID_IN_CWSCMS;
+import static gov.ca.cwds.service.messages.MessageCode.USER_WITH_EMAIL_EXISTS_IN_IDM;
+import static gov.ca.cwds.util.CurrentAuthenticatedUserUtil.getCurrentUser;
 import static gov.ca.cwds.util.Utils.toUpperCase;
 
 import com.amazonaws.services.cognitoidp.model.UserType;
 import gov.ca.cwds.UniversalUserToken;
+import gov.ca.cwds.data.persistence.auth.CwsOffice;
+import gov.ca.cwds.data.persistence.auth.StaffPerson;
 import gov.ca.cwds.idm.dto.User;
 import gov.ca.cwds.idm.dto.UserUpdate;
 import gov.ca.cwds.idm.service.MappingService;
+import gov.ca.cwds.idm.service.authorization.AuthorizationService;
 import gov.ca.cwds.idm.service.cognito.CognitoServiceFacade;
 import gov.ca.cwds.idm.service.cognito.util.CognitoUtils;
 import gov.ca.cwds.rest.api.domain.UserIdmValidationException;
+import gov.ca.cwds.rest.api.domain.auth.GovernmentEntityType;
 import gov.ca.cwds.service.CwsUserInfoService;
 import gov.ca.cwds.service.dto.CwsUserInfo;
+import gov.ca.cwds.service.messages.MessageCode;
 import gov.ca.cwds.service.messages.MessagesService;
 import java.util.Collection;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.Set;
+import org.apache.commons.lang3.StringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
-import org.springframework.util.StringUtils;
 
 @Service
 @Profile("idm")
 public class ValidationServiceImpl implements ValidationService {
 
   private UserByAdminRolesValidator userByAdminRolesValidator;
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(ValidationServiceImpl.class);
 
   private MappingService mappingService;
 
@@ -39,9 +57,42 @@ public class ValidationServiceImpl implements ValidationService {
 
   private CognitoServiceFacade cognitoServiceFacade;
 
-  @Override
-  public void validateUserCreate(UniversalUserToken admin, User user) {
+  private AuthorizationService authorizeService;
 
+  @Override
+  public User validateUserCreate(UniversalUserToken admin, User user) {
+    String racfId = user.getRacfid();
+    if (StringUtils.isNotBlank(racfId)) {
+      return validateRacfidUserCreate(user);
+    } else {
+        return user;
+    }
+  }
+
+  private User validateRacfidUserCreate(User user) {
+    String racfId = user.getRacfid();
+    String email = user.getEmail();
+
+    CwsUserInfo cwsUser = cwsUserInfoService.getCwsUserByRacfId(racfId);
+    if (cwsUser == null) {
+      throwValidationException(NO_USER_WITH_RACFID_IN_CWSCMS, racfId);
+    }
+
+    if (checkIfUserWithEmailExistsInCognito(email)) {
+      throwValidationException(USER_WITH_EMAIL_EXISTS_IN_IDM, email);
+    }
+
+    if (isActiveRacfIdPresent(racfId)) {
+      throwValidationException(ACTIVE_USER_WITH_RAFCID_EXISTS_IN_IDM, racfId);
+    }
+
+    enrichUserByCwsData(user, cwsUser);
+
+    Optional<MessageCode> authorizationError = buildAuthorizationError();
+    if (!authorizeService.canCreateUser(user) && authorizationError.isPresent()) {
+      throwValidationException(authorizationError.get(), user.getCountyName());
+    }
+    return user;
   }
 
   @Override
@@ -52,12 +103,63 @@ public class ValidationServiceImpl implements ValidationService {
     validateActivateUser(existedCognitoUser, updateUserDto);
   }
 
-  @Override
-  public boolean isActiveRacfIdPresent(String racfId) {
+  private void enrichUserByCwsData(User user, CwsUserInfo cwsUser) {
+    user.setRacfid(cwsUser.getRacfId());
+    enrichDataFromCwsOffice(cwsUser.getCwsOffice(), user);
+    enrichDataFromStaffPerson(cwsUser.getStaffPerson(), user);
+  }
+
+  private void enrichDataFromStaffPerson(StaffPerson staffPerson, final User user) {
+    if (staffPerson != null) {
+      user.setFirstName(staffPerson.getFirstName());
+      user.setLastName(staffPerson.getLastName());
+      user.setEndDate(staffPerson.getEndDate());
+      user.setStartDate(staffPerson.getStartDate());
+    }
+  }
+
+  private void enrichDataFromCwsOffice(CwsOffice office, final User user) {
+    if (office != null) {
+      user.setOfficeId(office.getOfficeId());
+      Optional.ofNullable(office.getPrimaryPhoneNumber())
+          .ifPresent(e -> user.setPhoneNumber(e.toString()));
+      Optional.ofNullable(office.getPrimaryPhoneExtensionNumber())
+          .ifPresent(user::setPhoneExtensionNumber);
+      Optional.ofNullable(office.getGovernmentEntityType())
+          .ifPresent(
+              x -> user.setCountyName((GovernmentEntityType.findBySysId(x)).getDescription()));
+    }
+  }
+
+  private Optional<MessageCode> buildAuthorizationError() {
+    switch (getStrongestAdminRole((getCurrentUser()))) {
+      case COUNTY_ADMIN:
+        return Optional.of(NOT_AUTHORIZED_TO_ADD_USER_FOR_OTHER_COUNTY);
+      case OFFICE_ADMIN:
+        return Optional.of(NOT_AUTHORIZED_TO_ADD_USER_FOR_OTHER_OFFICE);
+      default:
+        return Optional.empty();
+    }
+  }
+
+  private boolean isActiveRacfIdPresent(String racfId) {
     Collection<UserType> cognitoUsersByRacfId =
         cognitoServiceFacade.searchAllPages(composeToGetFirstPageByRacfId(toUpperCase(racfId)));
     return !CollectionUtils.isEmpty(cognitoUsersByRacfId)
         && isActiveUserPresent(cognitoUsersByRacfId);
+  }
+
+  private boolean checkIfUserWithEmailExistsInCognito(String email) {
+    Collection<UserType> cognitoUsers =
+        cognitoServiceFacade.searchPage(composeToGetFirstPageByEmail(email)).getUsers();
+    return !CollectionUtils.isEmpty(cognitoUsers);
+  }
+
+  private void throwValidationException(MessageCode messageCode, Object... args) {
+    String msg = messages.getTechMessage(messageCode, args);
+    String userMsg = messages.getUserMessage(messageCode, args);
+    LOGGER.error(msg);
+    throw new UserIdmValidationException(msg, userMsg, messageCode);
   }
 
   private User getNewUser(UserType existedCognitoUser, UserUpdate updateUserDto) {
@@ -87,10 +189,9 @@ public class ValidationServiceImpl implements ValidationService {
       return;
     }
     String racfId = CognitoUtils.getRACFId(existedCognitoUser);
-    if (StringUtils.isEmpty(racfId)) {
-      return;
+    if (StringUtils.isNotBlank(racfId)) {
+      validateActivateUser(racfId);
     }
-    validateActivateUser(racfId);
   }
 
   private static boolean canChangeToEnableActiveStatus(Boolean newEnabled, Boolean currentEnabled) {
@@ -145,5 +246,11 @@ public class ValidationServiceImpl implements ValidationService {
   public void setCognitoServiceFacade(
       CognitoServiceFacade cognitoServiceFacade) {
     this.cognitoServiceFacade = cognitoServiceFacade;
+  }
+
+  @Autowired
+  public void setAuthorizeService(
+      AuthorizationService authorizeService) {
+    this.authorizeService = authorizeService;
   }
 }
